[refactor] namespace structure
