+++ src/Connection_Oracle12.php
@@ -5,10 +5,8 @@ declare(strict_types=1);
 namespace atk4\dsql;
 
 /**
- * Custom Connection class specifically for Oracle 12c database.
+ * @deprecated use Oracle\Version12c\Connection instead - will be removed dec-2020
  */
-class Connection_Oracle12 extends Connection_Oracle
+class Connection_Oracle12 extends Oracle\Version12c\Connection
 {
-    /** @var string Query classname */
-    protected $query_class = Query_Oracle12c::class;
 }
