+++ src/Expression_MySQL.php
@@ -5,15 +5,8 @@ declare(strict_types=1);
 namespace atk4\dsql;
 
 /**
- * Perform query operation on MySQL server.
+ * @deprecated use MySQL\Expression instead - will be removed dec-2020
  */
-class Expression_MySQL extends Expression
+class Expression_MySQL extends MySQL\Expression
 {
-    /**
-     * Field, table and alias name escaping symbol.
-     * By SQL Standard it's double quote, but MySQL uses backtick.
-     *
-     * @var string
-     */
-    protected $escape_char = '`';
 }
