+++ src/PgSQL/Connection.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace atk4\dsql\PgSQL;
+
+use atk4\dsql\Connection as BaseConnection;
+
+/**
+ * Custom Connection class specifically for PostgreSQL database.
+ */
+class Connection extends BaseConnection
+{
+    /** @var string Query classname */
+    protected $query_class = Query::class;
+}
