+++ src/Query_Oracle.php
@@ -5,58 +5,8 @@ declare(strict_types=1);
 namespace atk4\dsql;
 
 /**
- * Perform query operation on Oracle server.
+ * @deprecated use Oracle\Query instead - will be removed dec-2020
  */
-class Query_Oracle extends Query_Oracle_Abstract
+class Query_Oracle extends Oracle\Query
 {
-    /**
-     * SELECT template.
-     *
-     * Default [limit] syntax not supported. Add rownum implementation instead.
-     *
-     * @var string
-     */
-    protected $template_select = 'select[option] [field] [from] [table][join][where][group][having][order]';
-    protected $template_select_limit = 'select * from (select rownum "__dsql_rownum","__t".* [from] (select[option] [field] [from] [table][join][where][group][having][order]) "__t") where "__dsql_rownum">[limit_start][and_limit_end]';
-
-    /**
-     * Limit how many rows will be returned.
-     *
-     * @param int $cnt   Number of rows to return
-     * @param int $shift Offset, how many rows to skip
-     *
-     * @return $this
-     */
-    public function limit($cnt, $shift = null)
-    {
-        // This is for pre- 12c version
-        $this->template_select = $this->template_select_limit;
-
-        return parent::limit($cnt, $shift);
-    }
-
-    /**
-     * Renders [limit_start].
-     *
-     * @return string rendered SQL chunk
-     */
-    public function _render_limit_start()
-    {
-        return (int) $this->args['limit']['shift'];
-    }
-
-    /**
-     * Renders [and_limit_end].
-     *
-     * @return string rendered SQL chunk
-     */
-    public function _render_and_limit_end()
-    {
-        if (!$this->args['limit']['cnt']) {
-            return '';
-        }
-
-        return ' and "__dsql_rownum"<=' .
-            ((int) ($this->args['limit']['cnt'] + $this->args['limit']['shift']));
-    }
 }
