+++ src/Query_Oracle12c.php
@@ -5,28 +5,8 @@ declare(strict_types=1);
 namespace atk4\dsql;
 
 /**
- * Perform query operation on Oracle server.
+ * @deprecated use Oracle\Version12c\Query instead - will be removed dec-2020
  */
-class Query_Oracle12c extends Query_Oracle_Abstract
+class Query_Oracle12c extends Oracle\Version12c\Query
 {
-    /**
-     * Renders [limit].
-     *
-     * @return string rendered SQL chunk
-     */
-    public function _render_limit()
-    {
-        if (isset($this->args['limit'])) {
-            $cnt = (int) $this->args['limit']['cnt'];
-            $shift = (int) $this->args['limit']['shift'];
-
-            return ' ' . trim(
-                ($shift ? 'OFFSET ' . $shift . ' ROWS' : '') .
-                ' ' .
-                // as per spec 'NEXT' is synonymous to 'FIRST', so not bothering with it.
-                // https://docs.oracle.com/javadb/10.8.3.0/ref/rrefsqljoffsetfetch.html
-                ($cnt ? 'FETCH NEXT ' . $cnt . ' ROWS ONLY' : '')
-            );
-        }
-    }
 }
