+++ src/Query_SQLite.php
@@ -5,32 +5,8 @@ declare(strict_types=1);
 namespace atk4\dsql;
 
 /**
- * Perform query operation on SQLite server.
+ * @deprecated use PgSQL\Query instead - will be removed dec-2020
  */
-class Query_SQLite extends Query
+class Query_SQLite extends SQLite\Query
 {
-    /**
-     * SQLite specific TRUNCATE template.
-     *
-     * @var string
-     */
-    protected $template_truncate = 'delete [from] [table_noalias]';
-
-    /**
-     * Returns a query for a function, which can be used as part of the GROUP
-     * query which would concatenate all matching fields.
-     *
-     * MySQL, SQLite - group_concat
-     * PostgreSQL - string_agg
-     * Oracle - listagg
-     *
-     * @param mixed  $field
-     * @param string $delimiter
-     *
-     * @return Expression
-     */
-    public function groupConcat($field, $delimeter = ',')
-    {
-        return $this->expr('group_concat({}, [])', [$field, $delimeter]);
-    }
 }
